requirement.txt
click==7.1.2
colorama==0.4.4
configparser==5.0.2
logzero==1.7.0
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
psutil==5.8.0
selenium==3.141.0
urllib3==1.26.4
